Make modification to send ARkit frames to phone app.
This doesn't really work because we're sending far too much data at
each frame, and the data's not well formated.
Ultimately we only want to send the ARKit model information, and not
the image captured by the camera, but for testing purposes this is
easier. This allows us to not need a projector for dev.

Also this is using the API given by peertalk, which certainly adds
overhead on top of TPC communication. We could try to use the TCP
layer directly. If we don't want to copy too much data around, that'll
help.

What's nice about using ARKit is that we get _all_ the math done for
us. ARKit does the detection and has a mode for flat surfaces, which
is an easier problem to solve than 3D tracking. This should give us
a 60fps framerate in terms of calculating the positions of everything.

We're not really interested in the RealTalk OS part, more in the
challenge of representing programs with physical objects and doing
low latency tracking + rendering. Low latency / high framerate
projectors are expensive (> $400) so we're not sure what to do here
apart from test with cheaper hardware and see how far we can push it.
Also interested in solving the issue that Dynamicland has with having
the dots (barcode) being covered by people's hands and arms, making
them stop running. I believe a bit of smart algorithm could mitigate
this issue (differentiate moving the paper and an arm obstructing the
camera, thanks to depth sensing part of ARKit).
If that's still bad we could use a glass table with QR codes under the
papers and a camera below the table.